feat(api): move controller route serializer to BC
