+++ api/src/prescription/target-profile/application/admin-target-profile-route.js
@@ -8,6 +8,62 @@ import { targetProfileController } from './admin-target-profile-controller.js';
 const register = async function (server) {
   server.route([
     {
+      method: 'PUT',
+      path: '/api/admin/target-profiles/{id}/outdate',
+      config: {
+        pre: [
+          {
+            method: (request, h) =>
+              securityPreHandlers.hasAtLeastOneAccessOf([
+                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
+                securityPreHandlers.checkAdminMemberHasRoleSupport,
+                securityPreHandlers.checkAdminMemberHasRoleMetier,
+              ])(request, h),
+            assign: 'hasAuthorizationToAccessAdminScope',
+          },
+        ],
+        validate: {
+          params: Joi.object({
+            id: identifiersType.targetProfileId,
+          }),
+        },
+        handler: targetProfileController.outdateTargetProfile,
+        tags: ['api', 'admin', 'target-profiles'],
+        notes: [
+          "- **Cette route est restreinte aux utilisateurs authentifiés ayant les droits d'accès**\n" +
+            '- Elle permet de marquer un profil cible comme obsolète',
+        ],
+      },
+    },
+    {
+      method: 'PUT',
+      path: '/api/admin/target-profiles/{id}/simplified-access',
+      config: {
+        pre: [
+          {
+            method: (request, h) =>
+              securityPreHandlers.hasAtLeastOneAccessOf([
+                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
+                securityPreHandlers.checkAdminMemberHasRoleSupport,
+                securityPreHandlers.checkAdminMemberHasRoleMetier,
+              ])(request, h),
+            assign: 'hasAuthorizationToAccessAdminScope',
+          },
+        ],
+        validate: {
+          params: Joi.object({
+            id: identifiersType.targetProfileId,
+          }),
+        },
+        handler: targetProfileController.markTargetProfileAsSimplifiedAccess,
+        tags: ['api', 'admin', 'target-profiles'],
+        notes: [
+          "- **Cette route est restreinte aux utilisateurs authentifiés ayant les droits d'accès**\n" +
+            '- Elle permet de marquer un profil cible comme étant "Parcours Accès Simplifié"',
+        ],
+      },
+    },
+    {
       method: 'POST',
       path: '/api/admin/target-profiles/{id}/attach-organizations',
       config: {
