+++ api/tests/prescription/target-profile/acceptance/application/admin-target-profile-route_test.js
@@ -357,4 +357,65 @@ describe('Acceptance | Route | admin-target-profile', function () {
       expect(response.statusCode).to.equal(200);
     });
   });
+
+  describe('PUT /api/admin/target-profiles/{id}/outdate', function () {
+    it('should return 204', async function () {
+      const targetProfile = databaseBuilder.factory.buildTargetProfile();
+      const user = databaseBuilder.factory.buildUser.withRole();
+      await databaseBuilder.commit();
+
+      const options = {
+        method: 'PUT',
+        url: `/api/admin/target-profiles/${targetProfile.id}/outdate`,
+        headers: { authorization: generateValidRequestAuthorizationHeader(user.id) },
+        payload: {
+          data: {
+            attributes: {
+              outdated: true,
+            },
+          },
+        },
+      };
+
+      // when
+      const response = await server.inject(options);
+
+      // then
+      expect(response.statusCode).to.equal(204);
+    });
+  });
+
+  describe('PUT /api/admin/target-profiles/{id}/simplified-access', function () {
+    it('should return 200 HTTP status code', async function () {
+      // given
+      const server = await createServer();
+
+      const user = databaseBuilder.factory.buildUser.withRole();
+      const targetProfile = databaseBuilder.factory.buildTargetProfile({ isSimplifiedAccess: false });
+      await databaseBuilder.commit();
+
+      const options = {
+        method: 'PUT',
+        url: `/api/admin/target-profiles/${targetProfile.id}/simplified-access`,
+        headers: {
+          authorization: generateValidRequestAuthorizationHeader(user.id),
+        },
+      };
+
+      // when
+      const response = await server.inject(options);
+
+      // then
+      expect(response.statusCode).to.equal(200);
+      expect(response.payload).to.equal(
+        JSON.stringify({
+          data: {
+            type: 'target-profiles',
+            id: targetProfile.id.toString(),
+            attributes: { 'is-simplified-access': true },
+          },
+        }),
+      );
+    });
+  });
 });
