+++ api/tests/prescription/target-profile/unit/application/admin-target-profile-controller_test.js
@@ -234,4 +234,36 @@ describe('Unit | Controller | admin-target-profile-controller', function () {
       expect(response.statusCode).to.equal(200);
     });
   });
+
+  describe('#outdateTargetProfile', function () {
+    let request;
+
+    beforeEach(function () {
+      sinon.stub(usecases, 'outdateTargetProfile');
+
+      request = {
+        params: {
+          id: 123,
+        },
+      };
+    });
+
+    context('successful case', function () {
+      it('should succeed', async function () {
+        // when
+        const response = await targetProfileController.outdateTargetProfile(request, hFake);
+
+        // then
+        expect(response.statusCode).to.equal(204);
+      });
+
+      it('should outdate target profile', async function () {
+        // when
+        await targetProfileController.outdateTargetProfile(request, hFake);
+
+        // then
+        expect(usecases.outdateTargetProfile).to.have.been.calledWithMatch({ id: 123 });
+      });
+    });
+  });
 });
