+++ api/tests/prescription/target-profile/unit/infrastructure/serializers/jsonapi/target-profile-serializer_test.js
@@ -1,6 +1,6 @@
-import { TargetProfile } from '../../../../../lib/domain/models/TargetProfile.js';
-import * as serializer from '../../../../../lib/infrastructure/serializers/jsonapi/target-profile-serializer.js';
-import { expect } from '../../../../test-helper.js';
+import { TargetProfile } from '../../../../../../../lib/domain/models/TargetProfile.js';
+import * as serializer from '../../../../../../../src/prescription/target-profile/infrastructure/serializers/jsonapi/target-profile-serializer.js';
+import { expect } from '../../../../../../test-helper.js';
 
 describe('Unit | Serializer | JSONAPI | target-profile-serializer', function () {
   describe('#serialize', function () {
