+++ api/tests/unit/application/target-profiles/index_test.js
@@ -656,155 +656,6 @@ describe('Unit | Application | Target Profiles | Routes', function () {
     });
   });
 
-  describe('PUT /api/admin/target-profiles/{id}/outdate', function () {
-    const method = 'PUT';
-    const url = '/api/admin/target-profiles/123/outdate';
-    const payload = {};
-
-    context('when user has role "SUPER_ADMIN", "SUPPORT" or "METIER"', function () {
-      it('should return a response with an HTTP status code 204', async function () {
-        // given
-        sinon
-          .stub(securityPreHandlers, 'hasAtLeastOneAccessOf')
-          .withArgs([
-            securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-            securityPreHandlers.checkAdminMemberHasRoleSupport,
-            securityPreHandlers.checkAdminMemberHasRoleMetier,
-          ])
-          .callsFake(() => (request, h) => h.response(true));
-        sinon
-          .stub(targetProfileController, 'outdateTargetProfile')
-          .callsFake((request, h) => h.response('ok').code(204));
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(moduleUnderTest);
-
-        // when
-        const { statusCode } = await httpTestServer.request(method, url, payload);
-
-        // then
-        expect(statusCode).to.equal(204);
-      });
-
-      context('when target profile ID is invalid', function () {
-        it('should reject request with HTTP code 400', async function () {
-          // given
-          const httpTestServer = new HttpTestServer();
-          await httpTestServer.register(moduleUnderTest);
-
-          // when
-          const { statusCode } = await httpTestServer.request(method, '/api/admin/target-profiles/azerty/outdate');
-
-          // then
-          expect(statusCode).to.equal(400);
-        });
-      });
-    });
-
-    context('when user has role "CERTIF"', function () {
-      it('should return a response with an HTTP status code 403', async function () {
-        // given
-        sinon
-          .stub(securityPreHandlers, 'hasAtLeastOneAccessOf')
-          .withArgs([
-            securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-            securityPreHandlers.checkAdminMemberHasRoleSupport,
-            securityPreHandlers.checkAdminMemberHasRoleMetier,
-          ])
-          .callsFake(
-            () => (request, h) =>
-              h
-                .response({ errors: new Error('forbidden') })
-                .code(403)
-                .takeover(),
-          );
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(moduleUnderTest);
-
-        // when
-        const { statusCode } = await httpTestServer.request(method, url, payload);
-
-        // then
-        expect(statusCode).to.equal(403);
-      });
-    });
-  });
-
-  describe('PUT /api/admin/target-profiles/{id}/simplified-access', function () {
-    const method = 'PUT';
-    const url = '/api/admin/target-profiles/123/simplified-access';
-    const payload = {};
-
-    context('when user has role "SUPER_ADMIN", "SUPPORT" or "METIER"', function () {
-      it('should return a response with an HTTP status code 200', async function () {
-        // given
-        sinon
-          .stub(securityPreHandlers, 'hasAtLeastOneAccessOf')
-          .withArgs([
-            securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-            securityPreHandlers.checkAdminMemberHasRoleSupport,
-            securityPreHandlers.checkAdminMemberHasRoleMetier,
-          ])
-          .callsFake(() => (request, h) => h.response(true));
-        sinon
-          .stub(targetProfileController, 'markTargetProfileAsSimplifiedAccess')
-          .callsFake((request, h) => h.response('ok').code(200));
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(moduleUnderTest);
-
-        // when
-        const { statusCode } = await httpTestServer.request(method, url, payload);
-
-        // then
-        expect(statusCode).to.equal(200);
-      });
-
-      context('when target profile ID is invalid', function () {
-        it('should reject request with HTTP code 400', async function () {
-          // given
-          const httpTestServer = new HttpTestServer();
-          await httpTestServer.register(moduleUnderTest);
-
-          // when
-          const { statusCode } = await httpTestServer.request(
-            method,
-            '/api/admin/target-profiles/azerty/simplified-access',
-          );
-
-          // then
-          expect(statusCode).to.equal(400);
-        });
-      });
-    });
-
-    context('when user has role "CERTIF"', function () {
-      it('should return a response with an HTTP status code 403', async function () {
-        // given
-        sinon
-          .stub(securityPreHandlers, 'hasAtLeastOneAccessOf')
-          .withArgs([
-            securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-            securityPreHandlers.checkAdminMemberHasRoleSupport,
-            securityPreHandlers.checkAdminMemberHasRoleMetier,
-          ])
-          .callsFake(
-            () => (request, h) =>
-              h
-                .response({ errors: new Error('forbidden') })
-                .code(403)
-                .takeover(),
-          );
-        const httpTestServer = new HttpTestServer();
-        await httpTestServer.register(moduleUnderTest);
-
-        // when
-        const { statusCode } = await httpTestServer.request(method, url, payload);
-
-        // then
-        expect(statusCode).to.equal(403);
-      });
-    });
-  });
-
   describe('POST /api/admin/target-profiles', function () {
     let validPayload;
